KrishiMitra-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.2.0
python-multipart==0.0.6
pydantic==2.5.3
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4
bcrypt==4.1.2
marshmallow==3.20.1
